Moved App to folder under src
